refactor: split api into modules
