+++ src/store/store.jsx
@@ -9,12 +9,15 @@ import uiStateReducer from "../state/data/uiStateSlice";
 import loginFormReducer from "../state/forms/loginFormSlice";
 import registerFormReducer from "../state/forms/registerFormSlice";
 import appointmentFormReducer from "../state/forms/appointmentFormSlice";
+import accidentFormReducer from "../state/forms/accidentsInsurance";
+import patientsIntakeFormReducer from "../state/forms/patientsIntakeFormSlice";
 
 import loadingReducer from "../state/ui/loadingSlice";
 import uiErrorsReducer from "../state/ui/errorsSlice";
 import notificationsReducer from "../state/ui/notificationsSlice";
 import modalsReducer from "../state/ui/modalsSlice";
-import { apiSlice } from "../services/api";
+import { authApi } from "../services/authApi";
+import { reportApi } from "../services/reportApi";
 
 const persistConfig = {
   key: "root",
@@ -33,6 +36,8 @@ const rootReducer = combineReducers({
     loginForm: loginFormReducer,
     registerForm: registerFormReducer,
     appointmentForm: appointmentFormReducer,
+    accidentForm: accidentFormReducer,
+    patientsIntakeForm: patientsIntakeFormReducer,
   }),
   ui: combineReducers({
     loading: loadingReducer,
@@ -40,7 +45,8 @@ const rootReducer = combineReducers({
     notifications: notificationsReducer,
     modals: modalsReducer,
   }),
-  [apiSlice.reducerPath]: apiSlice.reducer,
+  [authApi.reducerPath]: authApi.reducer,
+  [reportApi.reducerPath]: reportApi.reducer,
 });
 
 const persistedReducer = persistReducer(persistConfig, rootReducer);
@@ -50,8 +56,8 @@ export const store = configureStore({
   middleware: (getDefaultMiddleware) =>
     getDefaultMiddleware({
       serializableCheck: false,
-    }).concat(apiSlice.middleware),
-  
+    }).concat(authApi.middleware, reportApi.middleware),
+
 });
 
 export const persistor = persistStore(store);
